+++ changelog.md
@@ -1,5 +1,11 @@
 ## SilentRotate Changelog
 
+#### v0.9.2-beta (2021-xx-yy)
+
+- Changing modes no longer requires to be out-of-combat
+- Grounding Totem History detects when the shaman dies
+- Grounding Totem History detects when the group of the shaman is changed
+
 #### v0.9.1-beta (2021-10-30)
 
 - Added Blessing of Protection mode